sql injection docker-compose completed

+++ ctfd.py
@@ -5,6 +5,7 @@ import yaml
 import json
 from decouple import config
 import subprocess
+import os.path
 
 
 # Initialize ctfcli with the CTFD_TOKEN and CTFD_URL.
@@ -50,18 +51,25 @@ def docker_container(directory):
     chall = open(challenge, 'r')
     challenge_yml = yaml.load(chall, Loader=yaml.FullLoader)
     if 'exposeService' in challenge_yml:
-        #building the image based on the docker file
-        tag = strip_special_characters(challenge_yml['name'])
-        dir = directory.replace(' ', '\\ ') #we need the full path to the dockerfile and need to take care of spacing as well
-        os.system(f"docker build -t {tag} {dir}") #build the image tag
-        #run the image and start it using the latest build , desired port in deatach mode
-        public_port = challenge_yml['exposeService']['publicPort']
-        internal_port = challenge_yml['exposeService']['internalPort']
-        
-        #stop the previous image
-        os.system(f"docker stop {tag}")
-        os.system(f"docker rm {tag}")
-        os.system(f"docker run -d --name {tag} -p {public_port}:{internal_port} {tag}:latest")
+        #if we have a docker-compose, we will give the priority to it otherwise we will
+        #use the dockerfile
+        if os.path.exist(f"{directory}/docker-compose.yml"):
+            #stop the old docker and spin up the new one
+            #this run in deatach mode
+            os.system("ocker compose down --remove-orphans -v && docker-compose up -d")
+        else:
+            #building the image based on the docker file
+            tag = strip_special_characters(challenge_yml['name'])
+            dir = directory.replace(' ', '\\ ') #we need the full path to the dockerfile and need to take care of spacing as well
+            os.system(f"docker build -t {tag} {dir}") #build the image tag
+            #run the image and start it using the latest build , desired port in deatach mode
+            public_port = challenge_yml['exposeService']['publicPort']
+            internal_port = challenge_yml['exposeService']['internalPort']
+            
+            #stop the previous image
+            os.system(f"docker stop {tag}")
+            os.system(f"docker rm {tag}")
+            os.system(f"docker run -d --name {tag} -p {public_port}:{internal_port} {tag}:latest")
         
         #we want to update the challenge.yml with the port info
         protocol = challenge_yml['exposeService']['protocol']
